xdg-shell: Add edge constraints
An edge constraint is an complementery state to the tiled state, meaning
that it's not only tiled, but constrained in a way that it can't resize
in that direction.

This typically means that the constrained edge is tiled against a
monitor edge. An example configuration is two windows tiled next to each
other on a single monitor. Together they cover the whole work area.

The left window would have the following tiled and edge constraint
state:

  [ tiled_top, tiled_right, tiled_bottom, tiled_left,
    constrained_top, constrained_bottom, constrained_left ]

while the right window would have the following:

  [ tiled_top, tiled_right, tiled_bottom, tiled_left,
    constrained_top, constrained_bottom, constrained_right ]

This aims to replace and deprecate the `gtk_surface1.configure_edges`
event and the `gtk_surface1.edge_constraint` enum.

Signed-off-by: Jonas Ådahl <jadahl@gmail.com>