staging/ext-background-effect: fix capability value for blur
The value was zero because the enum was not a bitfield at some point when the protocol was
developed and I forgot to change the value to one when making it a bitfield.

This is technically a breaking change, but as the client could never receive the blur
capability before this commit, it won't actually break anything - it was already broken.

Signed-off-by: Xaver Hugl <xaver.hugl@kde.org>